+++ app/src/main/java/com/example/visit/database/TripsGet.java
@@ -1,6 +1,6 @@
 package com.example.visit.database;
 
-import com.example.visit.RecyclerViewItemMyTrips;
+import com.example.visit.travelling.RecyclerViewItemMyTrips;
 
 import java.util.ArrayList;
