+++ app/src/main/java/com/example/visit/planning/AccommodationFragment.java
@@ -1,4 +1,4 @@
-package com.example.visit;
+package com.example.visit.planning;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,21 +6,24 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.fragment.app.Fragment;
 
+import com.example.visit.MainActivity;
+import com.example.visit.R;
 import com.google.android.material.textfield.TextInputEditText;
 
 import java.util.Objects;
 
-public class TravellingNecessitiesFragment extends Fragment {
+public class AccommodationFragment extends Fragment {
 
-    public TravellingNecessitiesFragment() {
+    public AccommodationFragment() {
         // Required empty public constructor
     }
 
-    public static TravellingNecessitiesFragment newInstance() {
-        return new TravellingNecessitiesFragment();
+    public static AccommodationFragment newInstance() {
+        return new AccommodationFragment();
     }
 
     View view;
@@ -29,27 +32,29 @@ public class TravellingNecessitiesFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        view = inflater.inflate(R.layout.fragment_travelling_necessities, container, false);
+        view = inflater.inflate(R.layout.fragment_accommodation, container, false);
 
-        TextInputEditText necessitiesEdit = (TextInputEditText) view.findViewById(R.id.necessitiesEdit);
-        Button next = (Button) view.findViewById(R.id.next_necessities);
-        Button cancel = (Button) view.findViewById(R.id.cancel_necessities);
+        TextInputEditText accommodationEdit = (TextInputEditText) view.findViewById(R.id.accommodationEdit);
+        Button next = (Button) view.findViewById(R.id.next_accommodation);
+        Button cancel = (Button) view.findViewById(R.id.cancel_accommodation);
         TextView continueExploring = (TextView) view.findViewById(R.id.continue_text);
 
-        if (TripPlanning.getNecessities() != null) {
-            necessitiesEdit.setText(TripPlanning.getNecessities());
+        if (TripPlanning.getLocation() != null) {
+            accommodationEdit.setText(TripPlanning.getLocation());
         }
 
         next.setOnClickListener(view -> {
             //destination string holds the city user picked
-            String necessities = Objects.requireNonNull(necessitiesEdit.getText()).toString();
-            //  if user writes something in textbox, save it, but it is not required
-            if (necessities.length() > 0) {
-                TripPlanning.setNecessities(necessities);
-            }
+            String accommodation = Objects.requireNonNull(accommodationEdit.getText()).toString();
+            if (accommodation.length() > 0) {
+                TripPlanning.setLocation(accommodation);
 
-            MainActivity.changeFragment(Objects.requireNonNull(getActivity()).getSupportFragmentManager(), new ParticipantsFragment(), true);
+                MainActivity.changeFragment(Objects.requireNonNull(getActivity()).getSupportFragmentManager(), new TravellingModeFragment(), true);
+            } else {
+                Toast.makeText(view.getContext(), "Input accommodation!", Toast.LENGTH_LONG).show();
+            }
         });
+
         cancel.setOnClickListener(view -> {
             MainActivity.changeFragment(Objects.requireNonNull(getActivity()).getSupportFragmentManager(), new TripPlannerFragment(), true);
         });
@@ -66,4 +71,5 @@ public class TravellingNecessitiesFragment extends Fragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
     }
+
 }
