+++ app/src/main/java/com/example/visit/planning/ChosenTrip.java
@@ -1,4 +1,6 @@
-package com.example.visit;
+package com.example.visit.planning;
+
+import com.example.visit.travelling.RecyclerViewItemMyTrips;
 
 public class ChosenTrip {
