+++ app/src/main/java/com/example/visit/travelling/TravellingTimeFragment.java
@@ -1,4 +1,4 @@
-package com.example.visit;
+package com.example.visit.travelling;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,6 +8,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.fragment.app.Fragment;
+
+import com.example.visit.R;
 
 import org.json.JSONException;
 import org.json.JSONObject;
