+++ app/src/main/java/com/example/visit/travelling/TripDetailsFragment.java
@@ -1,4 +1,4 @@
-package com.example.visit;
+package com.example.visit.travelling;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -11,9 +11,13 @@ import android.widget.Toast;
 
 import androidx.fragment.app.Fragment;
 
+import com.example.visit.user_profile.LoggedUser;
+import com.example.visit.MainActivity;
+import com.example.visit.R;
 import com.example.visit.database.CurrentTripPatch;
 import com.example.visit.database.Database;
 import com.example.visit.database.HerokuAPI;
+import com.example.visit.planning.ChosenTrip;
 
 import org.jetbrains.annotations.NotNull;
