+++ app/src/main/res/layout/fragment_change_password.xml
@@ -9,7 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        tools:context=".ChangePasswordFragment">
+        tools:context=".user_profile.ChangePasswordFragment">
 
         <RelativeLayout
             android:id="@+id/changePasswordHeader"
