+++ app/src/main/res/layout/fragment_login.xml
@@ -16,7 +16,7 @@
         android:paddingTop="@dimen/fragment_vertical_margin"
         android:paddingRight="@dimen/fragment_horizontal_margin"
         android:paddingBottom="@dimen/fragment_vertical_margin"
-        tools:context=".LoginFragment">
+        tools:context=".user_profile.LoginFragment">
 
         <pl.droidsonroids.gif.GifImageView
             android:id="@+id/loginFragmentLoading"
