+++ app/src/main/res/layout/fragment_register.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@drawable/login_background"
-        tools:context=".RegisterFragment">
+        tools:context=".user_profile.RegisterFragment">
 
         <pl.droidsonroids.gif.GifImageView
             android:id="@+id/registerFragmentLoading"
