+++ app/src/main/res/layout/fragment_travelling.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/fragment_background_color"
-    tools:context=".TravellingFragment">
+    tools:context=".travelling.TravellingFragment">
 
 
     <!-- ScrollView's layout_height is equivalent to MATCH_CONSTRAINT
